Maybe Final? PCB Design
Put everything back on a single board for simplicity after considering that all the connectors between boards would probably make a pretty bad rats nest. Immediately had regrets about this decision after seeing how densely packed the board would have to be to fit within the constraints of Alnis's 1u frame, but pushed onward with design. Also a few of the expansion connectors were removed for size concerns, but other than that I think the schematic is the same as before.

This design just barely fits within the size constraints, but I am worried about some very long signal traces and if the 3mm power traces and the very segmented ground plane will be able to handle the high current. I attempted to stich together some parts of the plane on the front side, but it still isn't ideal.